Create Java 2D Array
Task:
You are given a  2D array. An hourglass in an array is a portion shaped like this:
a b c
  d
e f g

The sum of an hourglass is the sum of all the numbers within it. The sum for the hourglasses above are 7, 4, and 2, respectively.

In this problem you have to print the largest sum among all the hourglasses in the array.

Input Format
There will be exactly  lines, each containing  integers seperated by spaces. Each integer will be between  and  inclusive.

Output Format
Print the answer to this problem on a single line.