Create Java SubArray
Question
We define the following:
A subarray of an -element array is an array composed from a contiguous block of the original array's
elements. For example, if array = [1,2,3], then the subarrays are [1],[2] ,[3] ,[1,2] ,[2,3], and
[1,2,3]. Something like would not be a subarray as it's not a contiguous subsection of the original
array.
The sum of an array is the total sum of its elements.
An array's sum is negative if the total sum of its elements is negative.
An array's sum is positive if the total sum of its elements is positive.
Given an array of integers, find and print its number of negative subarrays on a new line.

Input Format
The first line contains a single integer,n, denoting the length of array A=[a0,a1....,an-1].
The second line contains space-separated integers describing each respective element,ai, in array A.

Constraints
1<= n <=100
-10^4 <= ai <= 10^4

Output Format
Print the number of subarrays of A having negative sums.